fix(systray): Simplify window visibility toggling
To hide the window, the previous approach would first check if the
window was:

a) Not already hidden.
b) Minimized.
c) The app was 'inactive'.

That approach was demonstrated to be problematic on Linux desktop
environments like GNOME 43, where the app could enter the
'inactive state' as soon as its window would lose focus, causing the
"Hide Notes" system tray action to attempt to show the window instead
of hide it.

This current approach is more straightforward: If the window is
currently hidden, we show it. If it's visible, we hide it.

I've also renamed a member variable to a more accurate name.