chore(ci): Switch to a better GitHub action to download build artifacts
The official GitHub action[1] to download build artifacts from previous
jobs, does not allow us to specify which files to download, so it ends
up downloading everything, including binaries from the macOS/Linux
workflows, which is both unnecessary and very slow.

This new action[2] is much more flexible and allows us to filter which
artifacts to download by their names.

[1] https://github.com/marketplace/actions/download-a-build-artifact
[2] https://github.com/marketplace/actions/download-workflow-artifact