chore(ci): Adjust the name of build artifacts
GitHub sorts the list of build artifacts alphabetically, so the first
thing I did was prefix the artifact's name with its operating system's
name. I think this makes more sense, at least visually.

Before/After:

- NotesSetup_2.2.1-Windows-release -> Windows-NotesSetup_2.2.1-release

The second minor thing I did was capitalize the value of the
'build-type' key, which is then forwarded to CMake's CMAKE_BUILD_TYPE,
and then used as a suffix of the artifact's name.

Before/After:

- NotesSetup_2.2.1-Windows-release -> Windows-NotesSetup_2.2.1-Release